chrome/pyproject.toml
[tool.poetry]
name = "chrome-focus"
version = "0.1.0"
description = "Chrome extension enforcement tool with daemon"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"
watchdog = "^3.0.0"
requests = "^2.31.0"
click = "^8.1.0"
psutil = "^5.9.0"
setproctitle = "^1.3.0"

[tool.poetry.scripts]
chrome-focus = "chrome_focus:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
